Moved configurations to one folder
